remove ControlsTrait trait
